+++ src/bin_evm_node/lib_dev/rpc_encodings.ml
@@ -208,20 +208,6 @@ module Kernel_version = MethodMaker (struct
   let method_ = "tez_kernelVersion"
 end)
 
-module Upgrade_nonce = MethodMaker (struct
-  type method_
-
-  type input = unit
-
-  type output = int32
-
-  let input_encoding = Data_encoding.unit
-
-  let output_encoding = Data_encoding.int32
-
-  let method_ = "tez_upgradeNonce"
-end)
-
 module Network_id = MethodMaker (struct
   type method_
 
@@ -667,7 +653,6 @@ end)
 let methods : (module METHOD) list =
   [
     (module Kernel_version);
-    (module Upgrade_nonce);
     (module Network_id);
     (module Chain_id);
     (module Accounts);