+++ src/kernel_evm/kernel/src/storage.rs
@@ -14,11 +14,10 @@ use tezos_evm_logging::{log, Level::*};
 use tezos_smart_rollup_core::MAX_FILE_CHUNK_SIZE;
 use tezos_smart_rollup_encoding::timestamp::Timestamp;
 use tezos_smart_rollup_host::path::*;
-use tezos_smart_rollup_host::runtime::{Runtime, RuntimeError, ValueType};
+use tezos_smart_rollup_host::runtime::{Runtime, ValueType};
 
 use crate::block_in_progress::BlockInProgress;
 use crate::error::{Error, StorageError};
-use crate::parsing::UPGRADE_NONCE_SIZE;
 use rlp::{Decodable, Encodable, Rlp};
 use tezos_ethereum::block::L2Block;
 use tezos_ethereum::rlp_helpers::FromRlpBytes;
@@ -664,18 +663,6 @@ pub fn store_kernel_upgrade_nonce<Host: Runtime>(
 ) -> Result<(), Error> {
     host.store_write_all(&KERNEL_UPGRADE_NONCE, &upgrade_nonce.to_le_bytes())
         .map_err(Error::from)
-}
-
-pub fn read_kernel_upgrade_nonce<Host: Runtime>(host: &mut Host) -> Result<u16, Error> {
-    match host.store_read_all(&KERNEL_UPGRADE_NONCE) {
-        Ok(bytes) => {
-            let slice_of_bytes: [u8; UPGRADE_NONCE_SIZE] =
-                bytes[..].try_into().map_err(|_| Error::InvalidConversion)?;
-            Ok(u16::from_le_bytes(slice_of_bytes))
-        }
-        Err(RuntimeError::PathNotFound) => Ok(1_u16),
-        Err(e) => Err(e.into()),
-    }
 }
 
 /// Get the index of accounts.